requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
python-multipart==0.0.6
Pillow==10.4.0
python-dotenv==1.0.0
bcrypt==4.1.2
passlib==1.7.4
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
tzdata
inference-sdk
pymysql
sqlalchemy
opencv-python-headless>=4.10.0
numpy>=1.24.0
